Fix cmd scheduler to check the availability of command bus Change initial value of m_timer to "max(bl, ccd_l)-1" for read command in c_BankStateRead.cpp

+++ src/sst/elements/CramSim/c_DeviceDriver.hpp
@@ -159,7 +159,7 @@ private:
 	Output *output;
 	DEBUG_MASK m_debugMask;
 	std::string m_debugPrefix;
-	unsigned m_issued_cmd;
+	uint64_t m_issued_cmd;
 };
 }
 }
